# remote-write: group data into batches to be sent over shards 