@bboreham 2021-12-02
Decided to go with #9830 as it has fewer moving parts.